requirements.txt
requests>=2.31.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
ccxt>=4.3.0
python-telegram-bot>=21.0
hyperliquid-python-sdk>=0.9.0
eth-account>=0.11.0
feedparser>=6.0.0
